Adds element assignment to lists.
Statements like list[index] = lvalue are valid. If index is out of
bounds, the interpreter gently normalizes it, which may not be the
best behaviour. There's a static check to see if list is indeed a list,
but note that currently it fails on strings -- the compiler thinks
that string is a subtype of list, but it isn't, not really.