Add mount command.
Mount allows to 'mount' external (real-life) directories as in-game
dirs. It only works on subdirs of (RL) pwd (for the sake of our sanity).
It doesn't do much yet, but does track the changes.

This means that if I modify/add/delete a file while the game is running,
I'll be able to use the changes in the game.

Caveats: don't mount a dir in a mounted dir. The computer won't explode,
but the inner mounting is lost whenever the outer is refreshed. Yeah, it's
naive, but we can call it undefined behaviour.

Also, the automatic refresh is called merely by the string_to_path, and
nothing guarantees that every relevant part calls it, so it's a big fat
TODO to write getters/setters (inserters, deleters) for Dir.